refactor(agents): registry service surface + agents edge seam files
registry::descriptor() replaces the four inline built_in_registry()
scans (sessions create/startup, model-registry http, runtime facade);
agents_contract.rs holds the dep-less wire mappers; agents_errors.rs
holds the one From<AgentRuntimeError> for ApiError impl; the
ProblemResponse second error mechanism is retired from agents.rs.
