+++ anyharness/crates/anyharness-lib/src/api/http/mod.rs
@@ -1,6 +1,8 @@
 pub mod access;
 pub mod agent_auth_config;
 pub mod agents;
+pub mod agents_contract;
+pub mod agents_errors;
 pub mod agents_model_registry;
 pub mod auth;
 pub mod blocking;
